rename includes to match GR naming scheme
