+++ include/ieee802_11/chunks_to_symbols.h
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2013, 2016 Bastian Bloessl <bloessl@ccs-labs.org>
+ * Copyright (C) 2013 Bastian Bloessl <bloessl@ccs-labs.org>
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -14,26 +14,23 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-#ifndef INCLUDED_IEEE802_11_SYNC_LONG_H
-#define INCLUDED_IEEE802_11_SYNC_LONG_H
+#ifndef INCLUDED_IEEE802_11_CHUNKS_TO_SYMBOLS_H
+#define INCLUDED_IEEE802_11_CHUNKS_TO_SYMBOLS_H
 
-#include <ieee802-11/api.h>
-#include <gnuradio/block.h>
+#include <ieee802_11/api.h>
+#include <gnuradio/tagged_stream_block.h>
 
 namespace gr {
 namespace ieee802_11 {
 
-class IEEE802_11_API sync_long : virtual public block
+class IEEE802_11_API chunks_to_symbols : virtual public tagged_stream_block
 {
 public:
-
-	typedef boost::shared_ptr<sync_long> sptr;
-	static sptr make(unsigned int sync_length,
-			bool log = false, bool debug = false);
-
+        typedef boost::shared_ptr<chunks_to_symbols> sptr;
+        static sptr make();
 };
 
-}  // namespace ieee802_11
-}  // namespace gr
+} /* namespace ieee802_11 */
+} /* namespace gr */
 
-#endif /* INCLUDED_IEEE802_11_SYNC_LONG_H */
+#endif /* INCLUDED_IEEE802_11_CHUNKS_TO_SYMBOLS_H */
