+++ include/ieee802_11/mac.h
@@ -14,36 +14,26 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-#ifndef INCLUDED_IEEE802_11_MAPPER_H
-#define INCLUDED_IEEE802_11_MAPPER_H
+#ifndef INCLUDED_IEEE802_11_MAC_H
+#define INCLUDED_IEEE802_11_MAC_H
 
-#include <ieee802-11/api.h>
+#include <ieee802_11/api.h>
 #include <gnuradio/block.h>
-
-enum Encoding {
-	BPSK_1_2  = 0,
-	BPSK_3_4  = 1,
-	QPSK_1_2  = 2,
-	QPSK_3_4  = 3,
-	QAM16_1_2 = 4,
-	QAM16_3_4 = 5,
-	QAM64_2_3 = 6,
-	QAM64_3_4 = 7,
-};
 
 namespace gr {
 namespace ieee802_11 {
 
-class IEEE802_11_API mapper : virtual public block
+class IEEE802_11_API mac : virtual public block
 {
 public:
 
-	typedef boost::shared_ptr<mapper> sptr;
-	static sptr make(Encoding mcs, bool debug = false);
-	virtual void set_encoding(Encoding mcs) = 0;
+	typedef boost::shared_ptr<mac> sptr;
+	static sptr make(std::vector<uint8_t> src_mac,
+			std::vector<uint8_t> dst_mac,
+			std::vector<uint8_t> bss_mac);
 };
 
 }  // namespace ieee802_11
 }  // namespace gr
 
-#endif /* INCLUDED_IEEE802_11_MAPPER_H */
+#endif /* INCLUDED_IEEE802_11_MAC_H */
