+++ include/ieee802_11/mapper.h
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2013 Bastian Bloessl <bloessl@ccs-labs.org>
+ * Copyright (C) 2013, 2016 Bastian Bloessl <bloessl@ccs-labs.org>
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -14,23 +14,36 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-#ifndef INCLUDED_IEEE802_11_CHUNKS_TO_SYMBOLS_H
-#define INCLUDED_IEEE802_11_CHUNKS_TO_SYMBOLS_H
+#ifndef INCLUDED_IEEE802_11_MAPPER_H
+#define INCLUDED_IEEE802_11_MAPPER_H
 
-#include <ieee802-11/api.h>
-#include <gnuradio/tagged_stream_block.h>
+#include <ieee802_11/api.h>
+#include <gnuradio/block.h>
+
+enum Encoding {
+	BPSK_1_2  = 0,
+	BPSK_3_4  = 1,
+	QPSK_1_2  = 2,
+	QPSK_3_4  = 3,
+	QAM16_1_2 = 4,
+	QAM16_3_4 = 5,
+	QAM64_2_3 = 6,
+	QAM64_3_4 = 7,
+};
 
 namespace gr {
 namespace ieee802_11 {
 
-class IEEE802_11_API chunks_to_symbols : virtual public tagged_stream_block
+class IEEE802_11_API mapper : virtual public block
 {
 public:
-        typedef boost::shared_ptr<chunks_to_symbols> sptr;
-        static sptr make();
+
+	typedef boost::shared_ptr<mapper> sptr;
+	static sptr make(Encoding mcs, bool debug = false);
+	virtual void set_encoding(Encoding mcs) = 0;
 };
 
-} /* namespace ieee802_11 */
-} /* namespace gr */
+}  // namespace ieee802_11
+}  // namespace gr
 
-#endif /* INCLUDED_IEEE802_11_CHUNKS_TO_SYMBOLS_H */
+#endif /* INCLUDED_IEEE802_11_MAPPER_H */
