+++ include/ieee802_11/signal_field.h
@@ -14,25 +14,28 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-#ifndef INCLUDED_IEEE802_11_PARSE_MAC_H
-#define INCLUDED_IEEE802_11_PARSE_MAC_H
 
-#include <ieee802-11/api.h>
-#include <gnuradio/block.h>
+#ifndef INCLUDED_IEEE802_11_SIGNAL_FIELD_H
+#define INCLUDED_IEEE802_11_SIGNAL_FIELD_H
+
+#include <ieee802_11/api.h>
+#include <gnuradio/digital/packet_header_default.h>
 
 namespace gr {
 namespace ieee802_11 {
 
-class IEEE802_11_API parse_mac : virtual public block
+class IEEE802_11_API signal_field : virtual public digital::packet_header_default
 {
 public:
+	typedef boost::shared_ptr<signal_field> sptr;
+	static sptr make();
 
-	typedef boost::shared_ptr<parse_mac> sptr;
-	static sptr make(bool log = false, bool debug = false);
-
+protected:
+	signal_field();
 };
 
 } // namespace ieee802_11
 } // namespace gr
 
-#endif /* INCLUDED_IEEE802_11_PARSE_MAC_H */
+#endif /* INCLUDED_IEEE802_11_SIGNAL_FIELD_H */
+
