+++ include/ieee802_11/sync_long.h
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2013 Bastian Bloessl <bloessl@ccs-labs.org>
+ * Copyright (C) 2013, 2016 Bastian Bloessl <bloessl@ccs-labs.org>
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -14,27 +14,26 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-#ifndef @GUARD_NAME@
-#define @GUARD_NAME@
+#ifndef INCLUDED_IEEE802_11_SYNC_LONG_H
+#define INCLUDED_IEEE802_11_SYNC_LONG_H
 
-#include <ieee802-11/api.h>
-#include <gnuradio/sync_block.h>
+#include <ieee802_11/api.h>
+#include <gnuradio/block.h>
 
 namespace gr {
 namespace ieee802_11 {
 
-class IEEE802_11_API @NAME@ : virtual public sync_block
+class IEEE802_11_API sync_long : virtual public block
 {
-public: 
-	typedef boost::shared_ptr<@NAME@> sptr;
-	static sptr make(int length);
+public:
 
-	virtual int length() const = 0;
-	virtual void set_length(int length) = 0;
+	typedef boost::shared_ptr<sync_long> sptr;
+	static sptr make(unsigned int sync_length,
+			bool log = false, bool debug = false);
 
 };
 
-} /* namespace blocks */
-} /* namespace gr */
+}  // namespace ieee802_11
+}  // namespace gr
 
-#endif /* @GUARD_NAME@ */
+#endif /* INCLUDED_IEEE802_11_SYNC_LONG_H */
