+++ include/ieee802_11/sync_short.h
@@ -14,26 +14,25 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-#ifndef INCLUDED_IEEE802_11_MAC_H
-#define INCLUDED_IEEE802_11_MAC_H
+#ifndef INCLUDED_IEEE802_11_SYNC_SHORT_H
+#define INCLUDED_IEEE802_11_SYNC_SHORT_H
 
-#include <ieee802-11/api.h>
+#include <ieee802_11/api.h>
 #include <gnuradio/block.h>
 
 namespace gr {
 namespace ieee802_11 {
 
-class IEEE802_11_API mac : virtual public block
+class IEEE802_11_API sync_short : virtual public block
 {
 public:
 
-	typedef boost::shared_ptr<mac> sptr;
-	static sptr make(std::vector<uint8_t> src_mac,
-			std::vector<uint8_t> dst_mac,
-			std::vector<uint8_t> bss_mac);
+	typedef boost::shared_ptr<sync_short> sptr;
+	static sptr make(double threshold, unsigned int min_plateau, bool log = false, bool debug = false);
+
 };
 
 }  // namespace ieee802_11
 }  // namespace gr
 
-#endif /* INCLUDED_IEEE802_11_MAC_H */
+#endif /* INCLUDED_IEEE802_11_SYNC_SHORT_H */
