+++ lib/chunks_to_symbols_impl.h
@@ -17,8 +17,8 @@
 #ifndef INCLUDED_IEEE802_11_CHUNKS_TO_SYMBOLS_IMPL_H
 #define INCLUDED_IEEE802_11_CHUNKS_TO_SYMBOLS_IMPL_H
 
-#include <ieee802-11/chunks_to_symbols.h>
-#include <ieee802-11/constellations.h>
+#include <ieee802_11/chunks_to_symbols.h>
+#include <ieee802_11/constellations.h>
 
 namespace gr {
 namespace ieee802_11 {
