+++ lib/decode_mac.cc
@@ -14,7 +14,7 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-#include <ieee802-11/decode_mac.h>
+#include <ieee802_11/decode_mac.h>
 
 #include "utils.h"
 #include "viterbi_decoder/viterbi_decoder.h"
