+++ swig/ieee802_11_swig.i
@@ -24,17 +24,17 @@
 %include "ieee802_11_swig_doc.i"
 
 %{
-#include "ieee802-11/chunks_to_symbols.h"
-#include "ieee802-11/constellations.h"
-#include "ieee802-11/decode_mac.h"
-#include "ieee802-11/ether_encap.h"
-#include "ieee802-11/frame_equalizer.h"
-#include "ieee802-11/mac.h"
-#include "ieee802-11/mapper.h"
-#include "ieee802-11/parse_mac.h"
-#include "ieee802-11/signal_field.h"
-#include "ieee802-11/sync_long.h"
-#include "ieee802-11/sync_short.h"
+#include "ieee802_11/chunks_to_symbols.h"
+#include "ieee802_11/constellations.h"
+#include "ieee802_11/decode_mac.h"
+#include "ieee802_11/ether_encap.h"
+#include "ieee802_11/frame_equalizer.h"
+#include "ieee802_11/mac.h"
+#include "ieee802_11/mapper.h"
+#include "ieee802_11/parse_mac.h"
+#include "ieee802_11/signal_field.h"
+#include "ieee802_11/sync_long.h"
+#include "ieee802_11/sync_short.h"
 %}
 
 %include "gnuradio/digital/packet_header_default.h"
@@ -44,17 +44,17 @@
 %ignore gr::digital::constellation_16qam;
 %include "gnuradio/digital/constellation.h"
 
-%include "ieee802-11/chunks_to_symbols.h"
-%include "ieee802-11/constellations.h"
-%include "ieee802-11/decode_mac.h"
-%include "ieee802-11/ether_encap.h"
-%include "ieee802-11/frame_equalizer.h"
-%include "ieee802-11/mac.h"
-%include "ieee802-11/mapper.h"
-%include "ieee802-11/parse_mac.h"
-%include "ieee802-11/signal_field.h"
-%include "ieee802-11/sync_long.h"
-%include "ieee802-11/sync_short.h"
+%include "ieee802_11/chunks_to_symbols.h"
+%include "ieee802_11/constellations.h"
+%include "ieee802_11/decode_mac.h"
+%include "ieee802_11/ether_encap.h"
+%include "ieee802_11/frame_equalizer.h"
+%include "ieee802_11/mac.h"
+%include "ieee802_11/mapper.h"
+%include "ieee802_11/parse_mac.h"
+%include "ieee802_11/signal_field.h"
+%include "ieee802_11/sync_long.h"
+%include "ieee802_11/sync_short.h"
 
 GR_SWIG_BLOCK_MAGIC2(ieee802_11, chunks_to_symbols);
 GR_SWIG_BLOCK_MAGIC2(ieee802_11, decode_mac);
